+++ src/core/models/index.ts
@@ -8,6 +8,7 @@
 
 export * from './artboard';
 export * from './artboard-info';
+export * from './artboard-map';
 export * from './directory';
 export * from './manifest-parser';
 export * from './resources';