+++ src/xd2svg.ts
@@ -2,22 +2,22 @@ import { Options } from 'extract-zip';
 import { dirSync, SynchrounousResult } from 'tmp';
 import { promisify } from 'util';
 import { CliOptions } from './cli/models';
-import { Directory } from './core/models';
-import { injectHtml, optimizeSvg, proceedFile } from './core/xd2svg';
+import { injectHtml, optimizeSvg, proceedFile } from './core';
+import { ArtboardMap, Directory } from './core/models';
 
 const extract: (zipPath: string, opts: Options) => Promise<void> = promisify(require('extract-zip'));
 
-export async function xd2svg(input: string | Directory, options: CliOptions): Promise<string | string[]> {
+export async function xd2svg(input: string | Directory, options: CliOptions): Promise<string | ArtboardMap> {
   const directory: Directory = typeof input === 'string' ?
     await openFile(input)
     : input;
   const isHtml: boolean = options.format === 'html';
-  const svg: string | string[] = proceedFile(directory, options.single);
+  const svg: string | ArtboardMap = proceedFile(directory, options.single);
 
-  const optimizedSvg: string | string[] =
+  const optimizedSvg: string | ArtboardMap =
     typeof svg === 'string' ?
       await prepareAndOptimizeSvg(svg, isHtml)
-      : await Promise.all(svg.map((curSvg) => prepareAndOptimizeSvg(curSvg, isHtml)));
+      : await promiseAllObject(svg, isHtml);
 
   if (typeof input !== 'string' && input.removeCallback) input.removeCallback();
 
@@ -33,6 +33,17 @@ async function openFile(inputFile): Promise<SynchrounousResult> {
     });
 
   return directory;
+}
+
+async function promiseAllObject(svg: ArtboardMap, isHtml): Promise<ArtboardMap> {
+  const keys = Object.keys(svg);
+  const values = await Promise.all(Object.values(svg).map((value: string) => prepareAndOptimizeSvg(value, isHtml)));
+
+  return keys.reduce((obj, key, index) => {
+    obj[key] = values[index];
+
+    return obj;
+  }, {});
 }
 
 function prepareAndOptimizeSvg(svg: string, isHtml: boolean): Promise<string> {