+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changes
+- Node.js API's
+
+### Fixed
+- Error when try to set output with subfolders
+- Incorrect filenames when `single=false`
 
 ## [0.6.0]
 ### Added